Taking input form user
We are taking data from user, basically we are making an UI for the project which will take input on left side and show output on the right side.

So we have made some designing and now we are going to use context in the next step to pass the data.